+++ src/api/middleware/atleta.ts
@@ -1,32 +1,44 @@
 import { TAtleta } from "@/types/TAtleta";
 
-
 export const atletaToApi = (atleta: TAtleta): any => {
-    return {
-        id: atleta.id,
-        nome: atleta.name,
-        nascimento: atleta.birthdate,
-        sexo: atleta.sex === "male" ? "M" : "F",
-        peso: atleta.weight,
-        altura: atleta.height,
-        categoria: atleta.category ?? "",
-        faixa: atleta.belt,
-        foto: atleta.photo,
-        email: atleta.email
-    }
-}
+  return {
+    id: atleta.id,
+    nome: atleta.name,
+    nascimento: atleta.birthdate,
+    sexo: atleta.sex === "male" ? "M" : "F",
+    peso: atleta.weight,
+    altura: atleta.height,
+    categoria: atleta.category ?? "",
+    faixa: atleta.belt,
+    foto: atleta.photo,
+    email: atleta.email,
+  };
+};
 
 export const apiToAtleta = (atletaApi: any): TAtleta => {
-    return {
-        id: atletaApi.id,
-        name: atletaApi.nome,
-        birthdate: atletaApi.nascimento,
-        sex: atletaApi.sexo,
-        weight: atletaApi.peso,
-        height: atletaApi.altura,
-        category: atletaApi.categoria,
-        belt: atletaApi.faixa,
-        photo: atletaApi.foto,
-        email: atletaApi.email
-    }
-}
+  return {
+    id: atletaApi.id,
+    name: atletaApi.nome,
+    birthdate: atletaApi.nascimento,
+    sex: atletaApi.sexo,
+    weight: atletaApi.peso,
+    height: atletaApi.altura,
+    category: atletaApi.categoria,
+    belt: atletaApi.faixa,
+    photo: atletaApi.foto,
+    email: atletaApi.email,
+  };
+};
+
+export const atletaToApiPost = (atleta: any): any => {
+  return {
+    nome: atleta.name,
+    nascimento: atleta.birthdate,
+    sexo: atleta.sex === "male" ? "M" : "F",
+    peso: atleta.weight,
+    altura: atleta.height,
+    faixa: atleta.belt,
+    foto: atleta.avatarBase64,
+    email: atleta.email,
+  };
+};
