+++ src/enums/faixas.ts
@@ -1,15 +1,14 @@
-export const Faixas =
-    [
-        "branca",
-        "cinza",
-        "azulClaro",
-        "azulEscuro",
-        "amarela",
-        "laranja",
-        "verde",
-        "roxa",
-        "marrom",
-        "preta",
-        "coral",
-        "vermelha",
-      ] as const;
+export const Faixas = [
+  "branca",
+  "cinza",
+  "azulClaro",
+  "azulEscuro",
+  "amarela",
+  "laranja",
+  "verde",
+  "roxa",
+  "marrom",
+  "preta",
+  "coral",
+  "vermelha",
+] as const;
